// domain/osustats/enums/OsuStatsBest.enum.ts
export enum EOsuStatsBestTimeframe {
    Yesterday = "Yesterday",
    LastWeek = "LastWeek",
    LastMonth = "LastMonth",
}

export enum EOsuStatsBestSort {
    Accuracy = "Accuracy",
    Combo = "Combo",
    Date = "Date",
    LeaderboardPosition = "LeaderboardPosition",
    Misses = "Misses",
    PP = "PP",
    Score = "Score",
}